Remove domain role from barbican service user
This patch semoves the openstack_user resource with :grant_domain
action. A user is always created within a specific domain; such a
mmembership cannot be tacked on later. This resource gave the user the
role intended for their project for the domain (i.e., for the Default
domain instead of for the service project).

We add the domain_name attribute that creates the barbican user in the
desired domain. Note that this change needs a sufficiently recent
openstackclient cookbook -- otherwise the domain_name attribute is
ignored (which does not matter as long as the barbican user is to be
created in the Default domain).